scripts/requirements.txt
apify-client
supabase
python-dotenv
pandas
requests
